# LUCENE-9852: Make Hunspell thread-safe