--- a/lucene/analysis/common/src/test/org/apache/lucene/analysis/hunspell/TestPerformance.java
+++ b/lucene/analysis/common/src/test/org/apache/lucene/analysis/hunspell/TestPerformance.java
@@ -28,11 +28,15 @@ import java.nio.file.Paths;
 import java.text.ParseException;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
+import java.util.concurrent.Future;
 import java.util.concurrent.TimeUnit;
 import java.util.function.Consumer;
 import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 import org.apache.lucene.util.LuceneTestCase;
+import org.apache.lucene.util.NamedThreadFactory;
 import org.junit.Assume;
 import org.junit.AssumptionViolatedException;
 import org.junit.BeforeClass;
@@ -106,24 +110,53 @@ public class TestPerformance extends LuceneTestCase {
     Dictionary dictionary = loadDictionary(code);
 
     List<String> words = loadWords(code, wordCount, dictionary);
+    List<String> halfWords = words.subList(0, words.size() / 2);
 
     Stemmer stemmer = new Stemmer(dictionary);
     Hunspell speller = new Hunspell(dictionary, TimeoutPolicy.NO_TIMEOUT, () -> {});
-    measure(
-        "Stemming " + code,
-        blackHole -> {
-          for (String word : words) {
-            blackHole.accept(stemmer.stem(word));
-          }
-        });
-    measure(
-        "Spellchecking " + code,
-        blackHole -> {
-          for (String word : words) {
-            blackHole.accept(speller.spell(word));
-          }
-        });
+    int cpus = Runtime.getRuntime().availableProcessors();
+    ExecutorService executor =
+        Executors.newFixedThreadPool(cpus, new NamedThreadFactory("hunspellStemming-"));
+
+    try {
+      measure("Stemming " + code, blackHole -> stemWords(words, stemmer, blackHole));
+
+      measure(
+          "Multi-threaded stemming " + code,
+          blackHole -> {
+            List<Future<?>> futures = new ArrayList<>();
+            for (int i = 0; i < cpus; i++) {
+              Stemmer localStemmer = new Stemmer(dictionary);
+              futures.add(executor.submit(() -> stemWords(halfWords, localStemmer, blackHole)));
+            }
+            try {
+              for (Future<?> future : futures) {
+                future.get();
+              }
+            } catch (Exception e) {
+              throw new RuntimeException(e);
+            }
+          });
+
+      measure(
+          "Spellchecking " + code,
+          blackHole -> {
+            for (String word : words) {
+              blackHole.accept(speller.spell(word));
+            }
+          });
+    } finally {
+      executor.shutdown();
+      assertTrue(executor.awaitTermination(1, TimeUnit.MINUTES));
+    }
+
     System.out.println();
+  }
+
+  private void stemWords(List<String> words, Stemmer stemmer, Consumer<Object> blackHole) {
+    for (String word : words) {
+      blackHole.accept(stemmer.stem(word));
+    }
   }
 
   private void checkSuggestionPerformance(String code, int wordCount) throws Exception {