@rmuir 2021-03-18
I guess my question is why does the stemmer need to be threadsafe?

Stemmers in lucene aren't threadsafe, we use a threadlocal model for the analysis chain. So tokenizers, stemmers, etc are cached per-thread, and maintain some buffers to avoid creating tons of garbage.

e.g. the way the Analyzer class works, if you are indexing with 8 threads, is that you have 8 HunspellStemFilters, each one with its own HunspellStemmer, so there are no thread safety issues. Previously the idea is that only the "large" thing (Dictionary) needed to be threadsafe as we don't want to instantiate it all the time anyway.