@donnerpeter 2021-03-18
The stemmer itself doesn't very much need to be thread-safe, but I'd prefer `Hunspell` API to be so, and the latter uses the former internally.