@rmuir 2021-03-18
OK, given that it doesn't hurt the performance that much and introduces no synchronization, seems like a good tradeoff.